+++ torii/__init__.py
@@ -5,4 +5,9 @@ try:
 except ImportError: # :nocov:
 	__version__ = 'unknown'
 
+from .warnings import install_handler
+
+# Install the warnings handler
+install_handler()
+
 __all__ = ()